+++ addressbook-selenium-test/src/test/java/CreateGroupTest.java
@@ -1,73 +1,30 @@
-
-
-import java.util.regex.Pattern;
-import java.util.concurrent.TimeUnit;
-
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.annotations.*;
-import static org.testng.Assert.*;
-import org.openqa.selenium.*;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.support.ui.Select;
 
-public class CreateGroupTest {
-  private WebDriver driver;
+public class CreateGroupTest extends TestBase {
 
-
-  @BeforeClass(alwaysRun = true)
-  public void setUp() throws Exception {
-    driver = new ChromeDriver();
-   driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-  }
-
-  @Test
-  public void testCreateGroup() throws Exception {
-    driver.get("http://localhost/addressbook/");
-    driver.findElement(By.name("user")).clear();
-    driver.findElement(By.name("user")).sendKeys("admin");
-    driver.findElement(By.name("pass")).clear();
-    driver.findElement(By.name("pass")).sendKeys("secret");
-    driver.findElement(By.name("user")).click();
-    driver.findElement(By.name("pass")).click();
-    driver.findElement(By.xpath("//input[@value='Login']")).click();
-    driver.findElement(By.linkText("groups")).click();
-    driver.findElement(By.name("new")).click();
-    driver.findElement(By.name("group_name")).click();
-    driver.findElement(By.name("group_name")).clear();
-    driver.findElement(By.name("group_name")).sendKeys("Name");
-    driver.findElement(By.name("group_header")).click();
-    driver.findElement(By.name("group_header")).clear();
-    driver.findElement(By.name("group_header")).sendKeys("Header");
-    driver.findElement(By.name("group_footer")).click();
-    driver.findElement(By.name("group_footer")).clear();
-    driver.findElement(By.name("group_footer")).sendKeys("Footer");
-    driver.findElement(By.name("submit")).click();
-    driver.findElement(By.linkText("group page")).click();
-  }
-
-  @AfterClass(alwaysRun = true)
-  public void tearDown() throws Exception {
-    driver.quit();
-
-  }
-
-  private boolean isElementPresent(By by) {
-    try {
-      driver.findElement(by);
-      return true;
-    } catch (NoSuchElementException e) {
-      return false;
+    @Test(priority = 2)
+    public void testCreateGroupLongName() {
+        goToGroupsPage("groups");
+        createNewGroup();
+        fillGroupForm("Name", "Header", "Footer");
+        submitGroupCreating();
+        System.out.println("testCreateGroupLongName passed");
     }
-  }
 
-  private boolean isAlertPresent() {
-    try {
-      driver.switchTo().alert();
-      return true;
-    } catch (NoAlertPresentException e) {
-      return false;
+    @Test(priority = 1)
+    public void testCreateGroupShortName() {
+        goToGroupsPage("groups");
+        createNewGroup();
+        fillGroupForm("N", "H", "F");
+        submitGroupCreating();
+        System.out.println("testCreateGroupShortName passed");
     }
-  }
 
-
+    @Test(priority = 3, enabled = false)
+    public void testCreateGroupEmpty() {
+        goToGroupsPage("groups");
+        createNewGroup();
+        fillGroupForm("", "", "");
+        submitGroupCreating();
+    }
 }
