+++ addressbook-selenium-test/src/test/java/DeleteGroupTest.java
@@ -1,0 +1,14 @@
+/*
+import org.testng.annotations.Test;
+
+public class DeleteGroupTest extends TestBase {
+    @Test
+    public void groupTestDelition(String groups) {
+        goToGroupsPage(groups);
+        selectGroup();
+        deleteGroup();
+    }
+
+
+}
+*/
