+++ addressbook-selenium-test/src/test/java/TestBase.java
@@ -1,0 +1,162 @@
+import org.openqa.selenium.By;
+import org.openqa.selenium.NoAlertPresentException;
+import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.support.ui.Select;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+
+import java.util.concurrent.TimeUnit;
+
+public class TestBase {
+    private WebDriver driver;
+
+    @BeforeClass(alwaysRun = true)
+    public void setUp() throws Exception {
+        driver = new ChromeDriver();
+        driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+        openAddressBook();
+        login();
+    }
+
+    public void openAddressBook() {
+        driver.get("http://localhost/addressbook/");
+    }
+
+    public void login() {
+        driver.findElement(By.name("user")).click();
+        driver.findElement(By.name("user")).clear();
+        driver.findElement(By.name("user")).sendKeys("admin");
+        driver.findElement(By.name("pass")).clear();
+        driver.findElement(By.name("pass")).sendKeys("secret");
+        driver.findElement(By.xpath("//input[@value='Login']")).click();
+    }
+
+
+    public void goToGroupsPage(String groups) {
+        driver.findElement(By.linkText(groups)).click();
+    }
+
+
+    public void createNewGroup() {
+        driver.findElement(By.name("new")).click();
+    }
+
+    public void fillGroupForm(String name, String header, String footer) {
+        driver.findElement(By.name("group_name")).click();
+        driver.findElement(By.name("group_name")).clear();
+        driver.findElement(By.name("group_name")).sendKeys(name);
+        driver.findElement(By.name("group_header")).click();
+        driver.findElement(By.name("group_header")).clear();
+        driver.findElement(By.name("group_header")).sendKeys(header);
+        driver.findElement(By.name("group_footer")).click();
+        driver.findElement(By.name("group_footer")).clear();
+        driver.findElement(By.name("group_footer")).sendKeys(footer);
+    }
+
+    public void submitGroupCreating() {
+        driver.findElement(By.name("submit")).click();
+        goToGroupsPage("group page");
+    }
+
+    public void selectGroup() {
+        driver.findElement(By.name("selected[]")).click();
+    }
+
+    public void deleteGroup() {
+        driver.findElement(By.name("delete[]")).click();
+    }
+
+    public void selectGroupToAdding() {
+        driver.findElement(By.name("new_group")).click();
+        new Select(driver.findElement(By.name("new_group"))).selectByVisibleText("Name");
+        driver.findElement(By.name("new_group")).click();
+
+
+    }
+
+    public void submitContactAdding() {
+        driver.findElement(By.xpath("(//input[@name='submit'])[2]")).click();
+    }
+
+    public void fillContactForm() {
+        driver.findElement(By.name("firstname")).click();
+        driver.findElement(By.name("firstname")).clear();
+        driver.findElement(By.name("firstname")).sendKeys("John");
+        driver.findElement(By.name("middlename")).click();
+        driver.findElement(By.name("middlename")).clear();
+        driver.findElement(By.name("middlename")).sendKeys("Name");
+        driver.findElement(By.name("lastname")).click();
+        driver.findElement(By.name("lastname")).clear();
+        driver.findElement(By.name("lastname")).sendKeys("Smith");
+        driver.findElement(By.name("nickname")).click();
+        driver.findElement(By.name("nickname")).clear();
+        driver.findElement(By.name("nickname")).sendKeys("Jonny");
+        driver.findElement(By.name("title")).click();
+        driver.findElement(By.name("title")).clear();
+        driver.findElement(By.name("title")).sendKeys("Title");
+        driver.findElement(By.name("company")).click();
+        driver.findElement(By.name("company")).clear();
+        driver.findElement(By.name("company")).sendKeys("Company");
+        driver.findElement(By.name("address")).click();
+        driver.findElement(By.name("address")).clear();
+        driver.findElement(By.name("address")).sendKeys("Address");
+        driver.findElement(By.name("home")).click();
+        driver.findElement(By.name("home")).clear();
+        driver.findElement(By.name("home")).sendKeys("Home");
+        driver.findElement(By.name("mobile")).click();
+        driver.findElement(By.name("mobile")).clear();
+        driver.findElement(By.name("mobile")).sendKeys("Mobile");
+        driver.findElement(By.name("work")).click();
+        driver.findElement(By.name("work")).clear();
+        driver.findElement(By.name("work")).sendKeys("Work");
+        driver.findElement(By.name("fax")).click();
+        driver.findElement(By.name("fax")).clear();
+        driver.findElement(By.name("fax")).sendKeys("Fax");
+        driver.findElement(By.name("email")).click();
+        driver.findElement(By.name("email")).clear();
+        driver.findElement(By.name("email")).sendKeys("e-mail");
+        driver.findElement(By.name("homepage")).click();
+        driver.findElement(By.name("homepage")).clear();
+        driver.findElement(By.name("homepage")).sendKeys("facebook.jonny.ru");
+        driver.findElement(By.name("bday")).click();
+        new Select(driver.findElement(By.name("bday"))).selectByVisibleText("1");
+        driver.findElement(By.name("bday")).click();
+        driver.findElement(By.name("bmonth")).click();
+        new Select(driver.findElement(By.name("bmonth"))).selectByVisibleText("January");
+        driver.findElement(By.name("bmonth")).click();
+        driver.findElement(By.name("byear")).click();
+        driver.findElement(By.name("byear")).clear();
+        driver.findElement(By.name("byear")).sendKeys("1980");
+    }
+
+    public void addContact(By add_new) {
+        driver.findElement(add_new).click();
+    }
+
+    @AfterClass(alwaysRun = true)
+    public void tearDown() throws Exception {
+        driver.quit();
+    }
+
+    private boolean isElementPresent(By by) {
+        try {
+            driver.findElement(by);
+            return true;
+        } catch (NoSuchElementException e) {
+            return false;
+        }
+    }
+
+    private boolean isAlertPresent() {
+        try {
+            driver.switchTo().alert();
+            return true;
+        } catch (NoAlertPresentException e) {
+            return false;
+        }
+    }
+
+
+}
